Updates cancel return type, error handling, tests
When successful, `cancel` returns an AppSubscription object with the
details of the cancelled subscription.

When unsuccessful, `cancel` will throw a `BillingError` with the details
of the error, in the case of an underlying `GraphqlQueryError` from the
`query` as well as any `userErrors` returned from the `query` itself.

This commit also adds tests and rearranges the type exports.